+++ src/home/models/contact_page.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import, unicode_literals
 from django.db import models
-from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 from wagtail.admin.edit_handlers import FieldPanel, StreamFieldPanel
 from wagtail.core import blocks
 from wagtail.core.fields import StreamField, RichTextField