+++ src/involvement/models/role.py
@@ -5,7 +5,7 @@ from django.forms import CheckboxSelectMultiple
 from django.apps import apps
 from django.conf import settings
 from django.db import models
-from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 from wagtail.admin.edit_handlers import MultiFieldPanel, FieldPanel, \
     FieldRowPanel
 from utils.translation import TranslatedField
@@ -153,7 +153,7 @@ class Role(models.Model):
         elif is_fum(user):
             return ['board']
         elif is_board(user):
-            return ['presidium']
+            return ['presidium', 'group_leader', 'engaged']
         elif is_presidium(user):
             return ['group_leader', 'engaged']
         elif is_group_leader(user):
@@ -172,7 +172,7 @@ class Role(models.Model):
         elif is_fum(user):
             return ['board', 'presidium']
         elif is_board(user):
-            return ['presidium']
+            return ['presidium', 'group_leader', 'engaged']
         elif is_presidium(user):
             return ['group_leader', 'engaged']
         elif is_group_leader(user):