+++ src/involvement/views/position_edit_view.py
@@ -4,8 +4,8 @@ from involvement.rule_utils import is_super
 
 
 class PositionEditView(EditView):
-    def get_form(self, form_class=None):
-        form = super(PositionEditView, self).get_form(form_class=form_class)
+    def get_form(self):
+        form = super(PositionEditView, self).get_form()
         queryset = form.fields['role'].queryset
         if not is_super(self.request.user):
             queryset = Role.edit_role_types_of(self.request.user)