+++ src/involvement/views/role_edit_view.py
@@ -2,5 +2,5 @@ from wagtail.contrib.modeladmin.views import EditView
 
 
 class RoleEditView(EditView):
-    def get_form(self, form_class=None):
-        return super(RoleEditView, self).get_form(form_class=form_class)
+    def get_form(self):
+        return super(RoleEditView, self).get_form()