+++ src/moore/urls.py
@@ -11,6 +11,8 @@ from wagtail.documents import urls as wagtaildocs_urls
 
 from .api import api_router
 
+from members.views import member_check_api
+
 urlpatterns = [
 
     # Needs to be imported before wagtail urls
@@ -18,6 +20,7 @@ urlpatterns = [
 
     # Needs to be imported before wagtail admin
     url(r'', include('involvement.urls')),
+    path('member_check_api/', member_check_api, name='member_check_api'),
 
     url(r'^admin/', include(wagtailadmin_urls)),
     url(r'^documents/', include(wagtaildocs_urls)),